docs whitepaper cleanup
