+++ mei-tra-frontend/public/llms.txt
@@ -1,7 +1,7 @@
 # 明専トランプ (Mei-Tra)
 
 > 明専トランプ（Meitra）は、4人2チーム制のオンライン戦略トランプゲームです。
-> ブロー宣言、ジャックシステム、ネグリなど独自ルールを持つトリックテイキングゲームで、
+> ビット、ジャックシステム、ネグリなど独自ルールを持つトリックテイキングゲームで、
 > 明治専門学校の学生がイギリス留学から持ち帰ったのが発祥と言われています。
 > Meitraはこの伝統カードゲームをオンラインで完全再現したWebアプリです。
